requirements.txt
pytest
pytest-repeat
python-socketio